Update UUIDWorker.php
Fix #1
I have fixed the problem. I'm not sure if it makes sence that the backup-field will be removed in the same step. Maybe it would be better to remove the backup columns in a seperate step with a seperate button? Otherwise what is the sence og the backup field???